Rename FsmStateBaseTest for StateTest and add namespace
