--- Assets/Tests/FsmStateTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 30660de190edfdc42bf6b030ea759114
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
